cloudServer/cloudFunction/requirements.txt
# Function dependencies, for example:
# package>=version
#opencv-python==4.7.0.72
opencv-python-headless
numpy==1.24.3
google-cloud-storage
functions-framework
Flask
